[#26]create-game-loop
