+++ srcs/input/key.c
@@ -6,7 +6,7 @@
 /*   By: tshimizu <tshimizu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/16 11:16:32 by tshimizu          #+#    #+#             */
-/*   Updated: 2026/05/17 11:10:01 by tshimizu         ###   ########.fr       */
+/*   Updated: 2026/05/31 11:18:54 by tshimizu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
