+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: tshimizu <tshimizu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/06 22:14:36 by tshimizu          #+#    #+#             */
-/*   Updated: 2026/05/23 18:43:14 by tshimizu         ###   ########.fr       */
+/*   Updated: 2026/05/31 10:33:08 by tshimizu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,20 +51,6 @@ static t_game	*setup_game(int argc, char *argv[])
 		return (free_array(lines), free_game(game), NULL);
 	free_array(lines);
 	return (game);
-}
-
-static void	register_hooks(t_game *game)
-{
-	mlx_hook(game->mlx_win, KEY_PRESS, KEY_PRESS_MASK, key_press, game);
-	mlx_hook(game->mlx_win, KEY_RELEASE, KEY_RELEASE_MASK, key_release, game);
-	mlx_hook(game->mlx_win, CLOSE_WINDOW, CLOSE_WINDOW_MASK, close_window,
-		game);
-}
-
-static void	run_game(t_game *game)
-{
-	register_hooks(game);
-	mlx_loop(game->mlx);
 }
 
 int	main(int argc, char *argv[])
